New WatermarkSigned test focusing on the stack overflow question 76544095
How to add a watermark in a digitally signed pdf using iText library without making signature invalid
https://stackoverflow.com/questions/76612048/how-to-add-a-watermark-in-a-digitally-signed-pdf-using-itext-library-without-mak
"Parent node.pdf", stamp.png
https://github.com/nickosl50/itext_valid_sign

WatermarkLikeNickLazidis - As already mentioned by the OP, the output of their code has invalidated signature and time stamp. The reason is that the code changes the static page content which is a disallowed change to signed PDFs.

WatermarkImproved - In contrast this method applies a watermark in an annotation. Consequentially, the signature is not invalidated.