New VerifySignature test focusing on the stack overflow question 77537388
"Unknown PdfException" by iText7 when TimeStampToken is not present in the embedded Pkcs7 UnsignedAttributes section
https://stackoverflow.com/questions/77537388/unknown-pdfexception-by-itext7-when-timestamptoken-is-not-present-in-the-embed
output.pdf as "riccardo_output.pdf"
https://drive.google.com/file/d/1v0zBc-YW-3VmANTms33gSa8mu5QVrjDG/view

testRiccardoOutput - Indeed, there is a regression: If the signature container has unsigned attributes but not a timestamp attribute among them, the code runs into an exception due to a method call on a null object. It's just fairly seldom that a signature has unsigned attributes but not a timestamp one, so the error has not popped up before.